# openapi.json
{
  "info": {
    "description": "API de AztecApp Explorer.\n\n**El sobre.** Todas las respuestas viajan envueltas: `{\"success\": true, \"data\": {...}}`\nen el camino feliz y `{\"success\": false, \"error\": \"...\", \"details\": {...}}` cuando algo\nfalla. El cliente debe desenvolver `data`.\n\n**Acceso.** El producto vende un DESBLOQUEO ÚNICO de 15 USD que no caduca; no hay\nniveles ni suscripción. Los endpoints públicos aceptan token opcional: sin él —o con\nuno revocado— responden 200 con el contenido en versión teaser, no 401. Eso es\ndeliberado: la app tiene que poder pintar el candado.\n\n**Qué se paga y qué no.** La logística de cada sitio (tarifa de taquilla, horarios,\ncómo llegar, avisos de seguridad) viaja SIEMPRE: son datos del mundo real y esconderlos\nsolo manda al viajero a buscarlos fuera. Lo que va detrás del desbloqueo es lo que\nescribimos nosotros: descripciones, `whyVisit`, contexto histórico y el audio de los\ntours.\n\n**Idioma.** `Accept-Language` se compara por prefijo de dos letras (`es-MX` → `es`).\nEl idioma base es el inglés y vive en las columnas del modelo; el español entra por la\ntabla de traducciones. Si una cuenta tiene `preferredLocale`, manda sobre la cabecera.",
    "title": "AztecApp Explorer API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5001",
      "description": "Backend local"
    },
    {
      "url": "http://localhost:4010",
      "description": "Mock de Prism"
    }
  ],
...
  "paths": {
...
    "/api/places/": {
      "get": {
        "tags": [
          "places"
        ],
        "summary": "Listado de sitios (pestaña Explore)",
        "description": "Token opcional: sin él la respuesta llega en versión teaser, con 200. Con sesión, cada sitio trae `isSaved` resuelto en la misma consulta.",
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10,
              "maximum": 100
            }
          },
          {
            "name": "curation",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "must_see",
                "quick_stop"
              ]
            },
            "description": "Filtros de la barra de Explore. Un valor desconocido se ignora."
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "es-MX"
            },
            "description": "Se compara por prefijo de dos letras. Soportados: en, es."
          }
        ],
        "responses": {
          "200": {
            "description": "Sitios",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceListResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/places/{place_id}": {
      "get": {
        "tags": [
          "places"
        ],
        "summary": "Ficha de un sitio",
        "description": "Bloqueada: llegan metadata y logística, no el contenido. Desbloqueada: llega todo.",
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "place_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "es-MX"
            },
            "description": "Se compara por prefijo de dos letras. Soportados: en, es."
          }
        ],
        "responses": {
          "200": {
            "description": "Ficha",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceResponse"
                }
              }
            }
          },
          "404": {
            "description": "No existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tours/": {
      "get": {
        "tags": [
          "tours"
        ],
        "summary": "Listado de tours",
        "description": "Los bloqueados salen como teaser en vez de desaparecer: si no, nadie llega a ver qué ofrece pagar.",
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10,
              "maximum": 100
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "es-MX"
            },
            "description": "Se compara por prefijo de dos letras. Soportados: en, es."
          }
        ],
        "responses": {
          "200": {
            "description": "Tours",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourListResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tours/user/tours": {
      "get": {
        "tags": [
          "tours"
        ],
        "summary": "Mis tours, con su avance",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "es-MX"
            },
            "description": "Se compara por prefijo de dos letras. Soportados: en, es."
          }
        ],
        "responses": {
          "200": {
            "description": "Tours empezados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserTourListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Sin token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/tours/{tour_id}": {
      "get": {
        "tags": [
          "tours"
        ],
        "summary": "Ficha de un tour",
        "description": "`stops` solo viaja con el desbloqueo. El teaser conserva `stopsCount` e `includesAudio` para pintar el candado.",
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "tour_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "es-MX"
            },
            "description": "Se compara por prefijo de dos letras. Soportados: en, es."
          }
        ],
        "responses": {
          "200": {
            "description": "Ficha del tour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourResponse"
                }
              }
            }
          },
          "404": {
            "description": "No existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/historical/content/place/{place_id}": {
      "get": {
        "tags": [
          "historical"
        ],
        "summary": "Artículos ligados a un sitio",
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "place_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "example": "es-MX"
            },
            "description": "Se compara por prefijo de dos letras. Soportados: en, es."
          }
        ],
        "responses": {
          "200": {
            "description": "Artículos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentListResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "openapi": "3.0.3",
  "components": {
    "schemas": {
      "Location": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "neighborhood": {
            "type": "string",
            "nullable": true
          },
          "distanceKm": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "latitude",
          "longitude"
        ],
        "additionalProperties": false
      },
      "Badges": {
        "type": "object",
        "properties": {
          "freeEntry": {
            "type": "boolean"
          },
          "outdoor": {
            "type": "boolean"
          },
          "archaeological": {
            "type": "boolean"
          }
        },
        "required": [
          "archaeological",
          "freeEntry",
          "outdoor"
        ],
        "additionalProperties": false
      },
...
      "Place": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "tagline": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "placeType": {
            "type": "string",
            "nullable": true
          },
          "curation": {
            "type": "string",
            "enum": [
              "must_see",
              "quick_stop",
              null
            ],
            "description": "Catálogo editorial. Alimenta los filtros de Explore.",
            "nullable": true
          },
          "rating": {
            "type": "number",
            "description": "Valoración editorial nuestra, no media de usuarios.",
            "nullable": true
          },
          "historicalSignificance": {
            "type": "string",
            "nullable": true
          },
          "estimatedVisitDuration": {
            "type": "integer",
            "description": "Minutos. Ordena y filtra.",
            "nullable": true
          },
          "visitDurationText": {
            "type": "string",
            "description": "Lo que se muestra: '2-4 hours'. Un entero no dice eso.",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "badges": {
            "$ref": "#/components/schemas/Badges"
          },
          "contentAccess": {
            "$ref": "#/components/schemas/ContentAccess"
          },
          "nearbyServices": {
            "$ref": "#/components/schemas/NearbyServices"
          },
          "isSaved": {
            "type": "boolean",
            "description": "El corazón. null sin sesión: no es false, es que no aplica.",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "nullable": true
          },
          "openingHours": {
            "type": "string",
            "nullable": true
          },
          "howToGetThere": {
            "type": "string",
            "nullable": true
          },
          "entryFee": {
            "$ref": "#/components/schemas/EntryFee"
          },
          "safetyRecommendations": {
            "type": "string",
            "nullable": true
          },
          "whyVisit": {
            "type": "string",
            "nullable": true
          },
          "historicalContext": {
            "$ref": "#/components/schemas/HistoricalContext"
          },
          "contentVerifiedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "badges",
          "contentAccess",
          "id",
          "location",
          "name",
          "nearbyServices"
        ],
        "additionalProperties": false
      },
      "PlaceList": {
        "type": "object",
        "properties": {
          "places": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Place"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "count": {
            "type": "integer"
          }
        },
        "required": [
          "places"
        ],
        "additionalProperties": false
      },
...
      "Tour": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "isFree": {
            "type": "boolean",
            "nullable": true
          },
          "isLocked": {
            "type": "boolean",
            "nullable": true
          },
          "unlockedForViewer": {
            "type": "boolean"
          },
          "estimatedDuration": {
            "type": "integer",
            "nullable": true
          },
          "durationText": {
            "type": "string",
            "nullable": true
          },
          "difficultyLevel": {
            "type": "string",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "totalDistance": {
            "type": "number",
            "nullable": true
          },
          "hasEntryFees": {
            "type": "boolean"
          },
          "stopsCount": {
            "type": "integer"
          },
          "includesAudio": {
            "type": "boolean"
          },
          "rating": {
            "type": "number",
            "nullable": true
          },
          "statistics": {
            "$ref": "#/components/schemas/TourStatistics"
          },
          "createdAt": {
            "type": "string",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "nullable": true
          },
          "contentDescription": {
            "type": "string",
            "description": "El guion. Solo con el desbloqueo.",
            "nullable": true
          },
          "stops": {
            "type": "array",
            "description": "Las paradas. Solo con el desbloqueo: el teaser conserva stopsCount e includesAudio para poder pintar el candado.",
            "items": {
              "$ref": "#/components/schemas/TourStop"
            }
          },
          "progress": {
            "description": "Solo en GET /tours/user/tours. Es un bloque reducido: el avance completo lo devuelven /start, /progress y /complete.",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserTourProgress"
              }
            ]
          }
        },
        "required": [
          "id",
          "title"
        ],
        "additionalProperties": false
      },
      "TourList": {
        "type": "object",
        "properties": {
          "tours": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tour"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "count": {
            "type": "integer"
          }
        },
        "required": [
          "tours"
        ],
        "additionalProperties": false
      },
      "HistoricalContent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "contentType": {
            "type": "string",
            "nullable": true
          },
          "era": {
            "type": "string",
            "nullable": true
          },
          "topic": {
            "type": "string",
            "nullable": true
          },
          "dateDescription": {
            "type": "string",
            "nullable": true
          },
          "readingTimeMinutes": {
            "type": "integer",
            "nullable": true
          },
          "timelineId": {
            "type": "string",
            "nullable": true
          },
          "sortOrder": {
            "type": "integer",
            "nullable": true
          },
          "placeId": {
            "type": "string",
            "nullable": true
          },
          "isLocked": {
            "type": "boolean"
          },
          "unlockedForViewer": {
            "type": "boolean"
          },
          "isRead": {
            "type": "boolean",
            "description": "null sin sesión.",
            "nullable": true
          },
          "author": {
            "type": "string",
            "nullable": true
          },
          "verified": {
            "type": "boolean",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "nullable": true
          },
          "nextContentId": {
            "type": "string",
            "description": "El botón Next de la cronología. Solo en el detalle.",
            "nullable": true
          },
          "content": {
            "type": "string",
            "description": "El texto. Solo desbloqueado."
          },
          "audioUrl": {
            "type": "string"
          },
          "videoUrl": {
            "type": "string"
          },
          "sources": {}
        },
        "required": [
          "id",
          "isLocked",
          "title",
          "unlockedForViewer"
        ],
        "additionalProperties": false
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "string"
          },
          "details": {
            "description": "Errores por campo cuando la validación falla."
          }
        },
        "required": [
          "error",
          "success"
        ],
        "additionalProperties": false
      },
      "PlaceResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/Place"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
      "PlaceListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/PlaceList"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
      "TourResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/Tour"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
      "TourListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/TourList"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
...
      "ContentListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "type": "object",
            "properties": {
              "content": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/HistoricalContent"
                }
              },
              "count": {
                "type": "integer"
              }
            },
            "required": [
              "content"
            ]
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
...
      "UserTourListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "type": "object",
            "properties": {
              "tours": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Tour"
                }
              },
              "count": {
                "type": "integer"
              }
            },
            "required": [
              "tours"
            ]
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}